+++ hip_data_tools/apache/kafka.py
@@ -15,13 +15,12 @@ from datetime import datetime
 import pandas as pd
 from confluent_kafka import Producer, Consumer
 
-from hip_data_tools.aws.s3 import S3Util
-from hip_data_tools.common import get_from_env_or_default_with_warning
-
 # TODO: For the kafka consumer implement the commit of the messages only after
 #   successful upload to s3. This will reduce the risks of data loss if the pod
 #   has errors after reading the data from Kafka
-from hip_data_tools.connect.aws import AwsConnectionManager, AwsConnectionSettings
+from hip_data_tools.aws.aws import AwsConnectionManager, AwsConnectionSettings
+from hip_data_tools.aws.s3 import S3Util
+from hip_data_tools.common import get_from_env_or_default_with_warning
 
 DEFAULT_PRODUCER_CONF = \
     """{'queue.buffering.max.messages': 10000,
