+++ hip_data_tools/common.py
@@ -5,6 +5,7 @@ Module contains variables and methods used for common / shared operations throug
 import logging
 import os
 import uuid
+from abc import ABC, abstractmethod
 
 LOG = logging.getLogger(__name__)
 """
@@ -62,3 +63,129 @@ def get_from_env_or_default_with_warning(env_var, default_val):
         value = default_val
 
     return value
+
+
+class KeyValueSource(ABC):
+    """
+    Abstract class for sourcing secrets, it is a key value source for retrieving values for keys
+    """
+
+    @abstractmethod
+    def get(self, key: str) -> str:
+        """
+        Abstract method to get the value for a given key
+        Args:
+            key (str): the key for which the value needs to be returned
+        Returns: str
+        """
+        pass
+
+    @abstractmethod
+    def exists(self, key):
+        """
+        Abstract methosd to verify if a key exists in the given data store
+        Args:
+            key (str): the key to be verified for existance
+        Returns: bool
+        """
+        pass
+
+
+class EnvironmentKeyValueSource(KeyValueSource):
+    """
+    class for sourcing secrets from env variables
+    """
+
+    def exists(self, key):
+        """
+        verify if a key exists
+        Args:
+            key (str): the key to be verified for existance
+        Returns: bool
+        """
+        if os.getenv(key, None) is None:
+            return False
+        return True
+
+    def get(self, key):
+        """
+        get the value for a given key
+        Args:
+            key (str): the key for which the value needs to be returned
+        Returns: str
+        """
+        return os.getenv(key)
+
+
+class DictKeyValueSource(KeyValueSource):
+    """
+    class for sourcing secrets from a provided Dict object, usually used for testing
+    """
+
+    def __init__(self, data):
+        self.data = data
+
+    def exists(self, key):
+        """
+        verify if a key exists
+        Args:
+            key (str): the key to be verified for existance
+        Returns: bool
+        """
+        if key in self.data:
+            return True
+        return False
+
+    def get(self, key):
+        """
+        get the value for a given key
+        Args:
+            key (str): the key for which the value needs to be returned
+        Returns: str
+        """
+        return self.data[key]
+
+
+ENVIRONMENT: EnvironmentKeyValueSource = EnvironmentKeyValueSource()
+
+
+class FileKeyValueSource(KeyValueSource):
+    """
+    class for sourcing secrets from file
+    """
+
+    def __init__(self, path):
+        pass
+
+    def exists(self, key):
+        pass
+
+    def get(self, key):
+        pass
+
+
+class SecretsManager(ABC):
+    """
+    A secret management abstract class that provides ways of extracting secrets
+    The class allows a subsequent connection class to use env vars to extract secrets in a
+    structured manner
+    Args:
+        required_keys (list[str]): a list of keys which will be checked for existence
+        source (KeyValueSource): a kv source that has secrets
+    """
+
+    def __init__(self, required_keys: list, source: KeyValueSource):
+        self.keys = required_keys
+        self._source = source
+        for key in self.keys:
+            if not self._source.exists(key):
+                raise Exception("Required Environment Variable {} does not exist!".format(key))
+
+    def get_secret(self, key):
+        """
+        get the secret valye for a given key
+        Args:
+            key (str): the key for given secret
+        Returns: str
+        """
+        return self._source.get(key)
